document vitessce code lifts
